converted the users module to use the action controllers
