+++ ACP3/Modules/ACP3/Users/Controller/Sidebar/Index/UserMenu.php
@@ -1,37 +1,22 @@
 <?php
-namespace ACP3\Modules\ACP3\Users\Controller\Sidebar;
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\Users\Controller\Sidebar\Index;
 
 use ACP3\Core;
 
 /**
- * Class Index
- * @package ACP3\Modules\ACP3\Users\Controller\Sidebar
+ * Class UserMenu
+ * @package ACP3\Modules\ACP3\Users\Controller\Sidebar\Index
  */
-class Index extends Core\Modules\Controller
+class UserMenu extends Core\Modules\Controller
 {
-    /**
-     * Displays the login mask, if the user is not already logged in
-     */
-    public function actionLogin()
-    {
-        if ($this->user->isAuthenticated() === false) {
-            $currentPage = base64_encode(($this->request->getArea() === 'admin' ? 'acp/' : '') . $this->request->getQuery());
-
-            $settings = $this->config->getSettings('users');
-
-            $this->view->assign('enable_registration', $settings['enable_registration']);
-            $this->view->assign('redirect_uri', $this->request->getPost()->get('redirect_uri', $currentPage));
-
-            $this->setTemplate('Users/Sidebar/index.login.tpl');
-        } else {
-            $this->setNoOutput(true);
-        }
-    }
-
     /**
      * Displays the user menu, if the user is logged in
      */
-    public function actionUserMenu()
+    public function execute()
     {
         if ($this->user->isAuthenticated() === true) {
             $userSidebar = [];
