+++ ACP3/Modules/ACP3/Users/Resources/config/controllers.yml
@@ -1,0 +1,139 @@
+parameters:
+  users.controller.admin.index.create.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Admin\Index\Create'
+  users.controller.admin.index.delete.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Admin\Index\Delete'
+  users.controller.admin.index.edit.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Admin\Index\Edit'
+  users.controller.admin.index.index.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Admin\Index\Index'
+  users.controller.admin.index.settings.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Admin\Index\Settings'
+  users.controller.frontend.account.edit.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Account\Edit'
+  users.controller.frontend.account.index.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Account\Index'
+  users.controller.frontend.account.settings.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Account\Settings'
+  users.controller.frontend.index.forgot_pwd.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Index\ForgotPwd'
+  users.controller.frontend.index.index.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Index\Index'
+  users.controller.frontend.index.login.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Index\Login'
+  users.controller.frontend.index.logout.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Index\Logout'
+  users.controller.frontend.index.register.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Index\Register'
+  users.controller.frontend.index.view_profile.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Frontend\Index\ViewProfile'
+  users.controller.sidebar.index.login.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Sidebar\Index\Login'
+  users.controller.sidebar.index.user_menu.fqdn: 'ACP3\Modules\ACP3\Users\Controller\Sidebar\Index\UserMenu'
+
+services:
+  users.controller.admin.index.create:
+    class: %users.controller.admin.index.create.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.date
+      - @core.helpers.formToken
+      - @core.helpers.secure
+      - @core.helpers.forms
+      - @users.model.userrepository
+      - @users.validator.admin
+      - @permissions.helpers
+
+  users.controller.admin.index.delete:
+    class: %users.controller.admin.index.delete.fqdn%
+    arguments:
+      - @core.context.admin
+      - @users.model.userrepository
+
+  users.controller.admin.index.edit:
+    class: %users.controller.admin.index.edit.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.helpers.formToken
+      - @core.helpers.secure
+      - @core.helpers.forms
+      - @users.model.userrepository
+      - @users.validator.admin
+      - @permissions.helpers
+
+  users.controller.admin.index.index:
+    class: %users.controller.admin.index.index.fqdn%
+    arguments:
+      - @core.context.admin
+      - @users.model.userrepository
+
+  users.controller.admin.index.settings:
+    class: %users.controller.admin.index.settings.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.helpers.formToken
+      - @core.helpers.forms
+      - @users.validation.admin_settings_form_validation
+
+  users.controller.frontend.account.edit:
+    class: %users.controller.frontend.account.edit.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @core.helpers.formToken
+      - @core.helpers.secure
+      - @users.model.userrepository
+      - @users.validator.account
+
+  users.controller.frontend.account.index:
+    class: %users.controller.frontend.account.index.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @users.model.userrepository
+
+  users.controller.frontend.account.settings:
+    class: %users.controller.frontend.account.settings.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @core.helpers.formToken
+      - @users.model.userrepository
+      - @user.validation.account_settings_form_validation
+
+  users.controller.frontend.index.forgot_pwd:
+    class: %users.controller.frontend.index.forgot_pwd.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @core.helpers.formToken
+      - @core.helpers.secure
+      - @users.model.userrepository
+      - @user.validation.account_forgot_password_form_validation
+      - @core.helpers.sendEmail
+    calls:
+      - ['setCaptchaHelpers', ['@?captcha.helpers']]
+
+  users.controller.frontend.index.index:
+    class: %users.controller.frontend.index.index.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @core.pagination
+      - @users.model.userrepository
+
+  users.controller.frontend.index.login:
+    class: %users.controller.frontend.index.login.fqdn%
+    parent: core.modules.controller.frontend
+
+  users.controller.frontend.index.logout:
+    class: %users.controller.frontend.index.logout.fqdn%
+    parent: core.modules.controller.frontend
+
+  users.controller.frontend.index.register:
+    class: %users.controller.frontend.index.register.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @core.date
+      - @core.helpers.formToken
+      - @core.helpers.secure
+      - @users.model.userrepository
+      - @users.validator.register
+      - @permissions.helpers
+      - @core.helpers.sendEmail
+    calls:
+      - ['setCaptchaHelpers', ['@?captcha.helpers']]
+
+  users.controller.frontend.index.view_profile:
+    class: %users.controller.frontend.index.view_profile.fqdn%
+    arguments:
+      - @core.context.frontend
+      - @users.model.userrepository
+
+  users.controller.sidebar.index.login:
+    class: %users.controller.sidebar.index.login.fqdn%
+    parent: core.modules.controller
+
+  users.controller.sidebar.index.user_menu:
+    class: %users.controller.sidebar.index.user_menu.fqdn%
+    parent: core.modules.controller
